Contact form completion
I have created a contact for that stores information in a my php database. if the input fields are incorrectly filled in when the form is submitted error messages will display and the the information will not be logged. If this occurs the input fields will remain populated. A success message is displayed when the form is submitted correctly.